move wicks to new file and improve prescreening
